requirements.txt
fastapi
uvicorn
pyttsx3
pillow
moviepy
